pyproject.toml
[tool.poetry]
name = "selenium-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.24.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.test.dependencies]
selenium = "^4.24.0"
pytest = "^8.3.3"
allure-pytest = "^2.13.5"


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.2"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150
